Lowering refactor (#72)
On behalf of @csarofeen 

Working on breaking up the lowering logic to be more incremental and easier to follow. This is in preparation to fix predicates after reductions and in combinations with following broadcasts.

This is to replace #65 for the updated base branch.

* Start commenting lowering better, break indexing pass into its own class.
* Refactor lowering to break up passes and make logic more incremental.
* removing commented code
