+++ torch/csrc/jit/codegen/cuda/ir_interface_nodes.h
@@ -275,10 +275,10 @@ struct TORCH_CUDA_API TensorView : public Val {
   }
 
   friend TORCH_CUDA_API TransformReplay;
-  // friend TORCH_CUDA_API TransformIter;
   friend TORCH_CUDA_API OptOutMutator;
-  friend TORCH_CUDA_API GPULower;
   friend TORCH_CUDA_API LoopNestGenerator;
+  friend void IRFixComputeAt(Fusion*);
+  friend void IRReplaceSizes();
 
  protected:
   // Make an exact copy of this tensor (similar to clone()), however, also grabs
