+++ torch/csrc/jit/codegen/cuda/lower_loops.h
@@ -8,69 +8,63 @@ namespace torch {
 namespace jit {
 namespace fuser {
 
-struct UnrollPass : public OptOutDispatch {
- private:
-  std::unordered_map<Expr*, Expr*> loop_replacement_map;
-  Fusion* fusion_;
-  const std::vector<Expr*>& incoming_exprs_;
-
-  // Keep all for loops conveniently to make unrolling easier
-  std::vector<ForLoop*> for_loops;
-
-  // keep track if we're within an unrolled loop
-  bool within_unroll = false;
-
-  // Custom dispatch for Expr, want to find out of it's a TV op
-  void handle(Expr*) final;
-
-  // Open the for loop.
-  void handle(ForLoop*) final;
-
-  UnrollPass(Fusion* _fusion, const std::vector<Expr*>& _incoming_exprs)
-      : fusion_(_fusion), incoming_exprs_(_incoming_exprs) {}
-
-  void computeMap();
-
- public:
-  static std::vector<Expr*> runPass(
-      Fusion* fusion,
-      const std::vector<Expr*>& exprs);
-};
-
+/*
+ * Loop nest generator pass will get IR that looks something like:
+ * T0[I0o{ceil(I0/4)}, I1o{ceil(I1/128)}, I0iU{4}, I1i{128}] = ...* for( i :
+ * I0o{ceil(I0/4)} ) { and will generate the loop nest structure for these exprs
+ * like:
+ *
+ * for( i : I0o{ceil(I0/4)} ) {
+ *   for( j : I1o{ceil(I1/128)} ) {
+ *     for( k : I0i{4} )
+ *       for( l : I1i{128} )
+ *         T0[I0o{ceil(I0/4)}, I1o{ceil(I1/128)}, I0iU{4}, I1i{128}] = ...
+ *
+ * It does not generate predicates, but it will generate allocations, and loop
+ * nests to initialize reduction buffers.
+ *
+ */
 struct TORCH_CUDA_API LoopNestGenerator : public OptOutDispatch {
  private:
+  // Lowered exprs to return
   std::vector<Expr*> lowered_exprs;
+  // Fusion pointer for convenience
   Fusion* fusion_;
 
-  // Keep all for loops conveniently to make unrolling easier
+  // Keep all for loops conveniently to make unrolling easier, basically just a
+  // stack of the active for_loops
   std::vector<ForLoop*> for_loops;
-  // computeAT scope is determined by the iterat domain, and the tensor view it
-  // belongs to (the final TensorView when following the computeAt path)
+
+  // Track the active computeAt scope, and what view we're "computeAt-ing" into
   std::vector<std::pair<IterDomain*, TensorView*>> compute_at_scope;
 
   // Create, place, and return the allocation for tv
   Expr* pushAlloc(TensorView*);
 
-  // Open a new inner most for loop
+  // Open a new inner most for loop, track which TV it was constructed from
+  // according to the computeAt chain.
   void openFor(std::pair<IterDomain*, TensorView*>);
+
+  // Close the inner most for loop
   void popFor();
 
   // Wrap pushBack in lower_utils if active_scope is null we want it to go
   // straight to lower_exprs
   void pushBack(Expr*);
 
-  // Update for loop structure based on this TensorView
+  // Update for loop structure based on this TensorView, see implementation for
+  // more details
   void updateLoopNest(TensorView*);
 
-  // Update for loop structure based on this TensorView, if there's an
-  // allocation stmt, send it in so we can make sure that we insert this
-  // initialization after it
+  // Initialize a buffer to init_val. If this buffer is in smem or registers,
+  // pass in its allocation statement so we can make sure that we insert this
+  // initialization comes after the allocation.
   void initReduction(TensorView* tv, Val* init_val, Expr* alloc_expr = nullptr);
 
-  // Check if a TV op, generate for loop nest around it
+  // Check if expr is a TV op and handle accordingly.
   void handle(Expr*) final;
 
-  // Generate the loop nest structure and place it in lowered_exprs
+  // Run the pass and accumulate output in lowered_exprs
   void generate(const std::vector<Expr*>& exprs);
 
   LoopNestGenerator(Fusion* _fusion) : fusion_(_fusion) {}
